se agrega servicio para eliminar notificaciones

+++ src/main/java/com/curiel/catalogos/repository/mercado/pago/NotificacionWebhooksRepository.java
@@ -3,6 +3,6 @@ package com.curiel.catalogos.repository.mercado.pago;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.curiel.catalogos.model.entity.mercado.pago.NotificacionWebhooks;
 
-public interface NotificacionWebhooksRepository extends JpaRepository<NotificacionWebhooks, Integer> {
+public interface NotificacionWebhooksRepository extends JpaRepository<NotificacionWebhooks, Long> {
 
 }
